rename explore_card to card3
