+++ 04-understanding-widgets/projects/starter/lib/home.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'card1.dart';
 import 'card2.dart';
-import 'explore_card.dart';
+import 'card3.dart';
 
 class Home extends StatefulWidget {
   const Home({Key key}) : super(key: key);
